requirements.txt
absl-py==2.1.0
braceexpand==0.1.7
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2024.6.0
ftfy==6.2.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.1
huggingface-hub==0.34.3
idna==3.7
Jinja2==3.1.6
Markdown==3.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==5.29.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.3
six==1.16.0
sympy==1.14.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
timm==1.0.19
tokenizers==0.21.4
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.54.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.5.0
wcwidth==0.2.13
webdataset==0.2.86
Werkzeug==3.0.6
